fix: rename entities
